pyproject.toml
[project]
name = "whatsapp-chat-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "loguru-config>=0.1.0",
    "requests>=2.32.3",
]
